♻️ Refactor: 컨벤션에 맞춰 폴더 구조 변경
