+++ src/app/sns-event-assistant/(default)/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 
-import { GnbSection, SnsGnbTabType } from '@common/constants/gnbs';
+import { GnbSection, SnsGnbTabType } from '@common/types/gnbs.types';
 
 import GnbLeftLayout from '@common/components/layouts/GnbLeftLayout.server';
