+++ src/common/components/gnbs/GnbLeft/NavItem/NavItem.types.ts
@@ -1,4 +1,4 @@
-import { GnbLeftSection } from '@common/constants/gnbs';
+import { GnbLeftSection } from '@common/types/gnbs.types';
 
 export interface NavItemProps {
   item: GnbLeftSection;
