+++ src/common/components/gnbs/GnbLeft/RecentWorkspace/RecentWorkspace.server.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-import { ROUTES } from '@common/constants/routes';
+import { ROUTES } from '@common/constants/routes.constants';
 
 import WorkspaceProfileImage from '@common/components/images/WorkspaceProfileImage/WorkspaceProfileImage.client';
