+++ src/common/constants/gnbs.constants.ts
@@ -1,18 +1,16 @@
 import { LeftGnbIconsState } from '@icons/LeftGnbIcons/LeftGnbIcons.types';
 
-import { ROUTES } from './routes';
+import {
+  GnbLeftSection,
+  GnbSection,
+  SectionConfigs,
+  SectionOptionKey,
+  SnsGnbTabType,
+} from '@common/types/gnbs.types';
 
-// 공통 gnb 섹션 타입
-export enum GnbSection {
-  MAIN = 'MAIN',
-  AI_MEETING_MANAGER = 'AI_MEETING_MANAGER',
-  SNS_EVENT_ASSISTANT = 'SNS_EVENT_ASSISTANT',
-  TEAM_MOOD_TRACKER = 'TEAM_MOOD_TRACKER',
-  CALENDAR = 'CALENDAR',
-  CUSTOM = 'CUSTOM',
-}
+import { ROUTES } from './routes.constants';
 
-// 공통 gnb 섹션 라벨
+// 공통 GNB 섹션 라벨
 export const GnbSectionLabels: Record<GnbSection, string> = {
   [GnbSection.MAIN]: '메인 홈',
   [GnbSection.AI_MEETING_MANAGER]: 'AI 회의 진행 매니저',
@@ -25,9 +23,6 @@ export const GnbSectionLabels: Record<GnbSection, string> = {
 // ─────────────────────────────────────────────
 
 // GnbLeft
-// LeftNav에 실제로 사용하는 값만 따로 필터링
-export type GnbLeftSection = Exclude<GnbSection, GnbSection.CUSTOM>;
-
 // nav 항목 리스트
 export const GnbLeftNavItems: GnbLeftSection[] = [
   GnbSection.MAIN,
@@ -58,16 +53,6 @@ export const GnbLeftNavItemIconState: Record<GnbLeftSection, LeftGnbIconsState> 
 // ─────────────────────────────────────────────
 
 // GnbTop
-// 옵션 키
-export enum SectionOptionKey {
-  SERVICE_HOME = 'SERVICE_HOME',
-  ALL_MEETINGS = 'ALL_MEETINGS',
-  ALL_EVENTS = 'ALL_EVENTS',
-  SNS_LINK_MANAGE = 'SNS_LINK_MANAGE',
-  ALL_REPORTS = 'ALL_REPORTS',
-  CALENDAR = 'CALENDAR',
-}
-
 // 옵션 키 라벨 매핑
 export const SectionOptionLabels: Record<SectionOptionKey, string> = {
   [SectionOptionKey.SERVICE_HOME]: '서비스 홈',
@@ -78,31 +63,9 @@ export const SectionOptionLabels: Record<SectionOptionKey, string> = {
   [SectionOptionKey.CALENDAR]: '캘린더',
 };
 
-// SNS 전용 탭 타입 및 라벨
-export enum SnsGnbTabType {
-  ALL_EVENTS = 'ALL_EVENTS',
-  SNS_LINK_MANAGE = 'SNS_LINK_MANAGE',
-}
-
 export const SnsGnbTabLabels: Record<SnsGnbTabType, string> = {
   [SnsGnbTabType.ALL_EVENTS]: '전체 이벤트',
   [SnsGnbTabType.SNS_LINK_MANAGE]: 'SNS 링크 관리',
-};
-
-// 섹션 구성 설정
-// 객체 형태
-type StaticSectionConfig = {
-  title: string;
-  options: { label: string; key: SectionOptionKey }[];
-};
-
-// 함수 형태 (CUSTOM용)
-type CustomSectionConfigFn = (title: string) => StaticSectionConfig;
-
-type SectionConfigs = {
-  [K in Exclude<GnbSection, GnbSection.CUSTOM>]: StaticSectionConfig;
-} & {
-  [GnbSection.CUSTOM]: CustomSectionConfigFn;
 };
 
 export const sectionConfigs: SectionConfigs = {
